refactor(mapgen): split authoring type ownership
